fix: move roles into users table
